pyproject.toml
[tool.poetry]
name = "yodapa"
version = "0.1.3"
description = "Personal Assistant on the command line"
authors = ["Man Parvesh Singh Randhawa <manparveshsinghrandhawa@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [
    { include = "yodapa", from = "src" }
]
repository = "https://github.com/yoda-pa/yoda/tree/v2"
homepage = "https://yoda-pa.github.io/"

[tool.poetry.dependencies]
python = "^3.9"
typer = "^0.12.5"
pytest = "^8.3.3"
pyyaml = "^6.0.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
yoda = "cli.yoda:app"

[tool.poetry.plugins."yodapa.plugins"]
hi_plugin = "yodapa.plugins.hi_plugin:HiPlugin"
